requirements/dev.txt
comet_ml==3.49.7
dvc==3.59.1
fire==0.7.0
gin-config==0.5.0
numpy==2.2.4
matplotlib==3.10.1
opencv-python-headless==4.11.0.86
pandas==2.2.3
scikit-image==0.25.2
scipy==1.15.2
skan==0.12.2
tox==4.24.1
